@@ -8,19 +8,19 @@ class OrganizationsController < ApplicationController
     @organizations = Organization.all
 
     respond_to do |format|
-      format.html # index.html.erb
+      format.html # index.html.slim
       format.json { render json: @organizations }
     end
   end
 
-  # GET /organizations/paperboy-all
-  # GET /organizations/paperboy-all.json
+  # GET /organizations/papslimoy-all
+  # GET /organizations/papslimoy-all.json
   def show
     @organization = Organization.find_by_login(params[:id])
     @snippets = @organization.snippets.published.page(params[:page]).decorate
 
     respond_to do |format|
-      format.html # show.html.erb
+      format.html # show.html.slim
       format.json { render json: @organization }
     end
   end
