@@ -1,7 +1,7 @@
 class StarsController < ApplicationController
   def create
     message = Message.find(params[:message_id])
-    star    = Star.find_or_create_by_user_id_and_message_id(current_user.id, message.id)
+    star    = Star.find_or_create_by(user_id: current_user.id, message_id: message.id)
 
     Star.transaction do
       star.increment(:count, 1)
