@@ -1,6 +1,6 @@
 Fabricator(:snippet) do
-  user         { Fabricate(:user) }
-  organization { Fabricate(:organization) }
+  user
+  organization
   title        { sequence(:title) {|i| "title#{i}" } }
   published    { true }
 end
